Move validation work to an external class; ValidationWorker
