+++ library/src/main/java/com/alexfu/formvalidator/ValidationWorker.java
@@ -1,0 +1,9 @@
+package com.alexfu.formvalidator;
+
+/**
+ * The actual worker that does the work of validating
+ */
+public interface ValidationWorker<T> {
+  T performValidate(ValidationTask[] tasks);
+  T performValidate(ValidationTask task);
+}
